reuse spreadsheet script
