--- data/festival/speakers.json
@@ -1,1 +1,0 @@
-[{"name":"Rachel Arnould","title":"Open Food","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":"OK"},{"name":"Dr Sonia Ben Mokhtar","title":"CNRS, LIRIS-DRIM","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Gilles Fedak","title":"iExec","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Pierre Boc","title":"Berty","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Antoine Champagne","title":"Reflets.info","picture":"https://pbs.twimg.com/profile_images/1253092436/k-pict_400x400.png","twitter":"https://twitter.com/_Kitetoa_","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Mathilde Imer","title":"Convention Citoyenne pour le Climat","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Alice Poggioli","title":"Digital Commons Consortium","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Boris Séguy","title":"Digital Commons Consortium","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Guillaume Rouyer","title":"Assemblée Virtuelle","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Fasto","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Lola Rigaut-Luczak","title":"","picture":"","twitter":"https://twitter.com/lola_dam","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Dani Shavit","title":"DAOincubator, AragonBlack","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Adrien Berthou","title":"Genesis DAO, FestDAO","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Alex Masmej","title":"MetaCartel DAO, Stake DAO","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Pietro Abate","title":"Tezos","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Alexandre Doussot","title":"Tezos","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Louis Giraux","title":"Aragon Association","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Danièle Bourcier","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Primavera de Filippi","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Lionel Maurel","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Axiom Team","title":"Monnaies Libres / Duniter","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Mamie Crypto","title":"@LesHackeuses","picture":"https://pbs.twimg.com/profile_images/925844484117626880/XrDfSqv8_400x400.jpg","twitter":"https://twitter.com/MamieCrypto","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Adli Takkal Bataille","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Jacques Favier","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Arthur Bard","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Fred Menez","title":"","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Manfred Touron","title":"Berty","picture":"https://pbs.twimg.com/profile_images/999637482420383744/PwcltvCI_400x400.jpg","twitter":"https://twitter.com/moul","linkedIn":"","github":"https://github.com/moul","mastodon":"","picture_valid":"OK"},{"name":"Antoine Eddi","title":"Berty","picture":"","twitter":"https://twitter.com/aeddi42","linkedIn":"","github":"https://github.com/aeddi","mastodon":"","picture_valid":""},{"name":"Thomas Chataignier","title":"Blockchain Partner","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Philippe M.","title":"Blockchain Partner","picture":"","twitter":"","linkedIn":"","github":"no pic","mastodon":"","picture_valid":""},{"name":"FX Thoorens","title":"CEO & Founder, ARK.io","picture":"","twitter":"https://twitter.com/fxthoorens","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Caleb James Delisle","title":"CJDNS","picture":"https://m.trnsz.com:8448/_matrix/media/r0/thumbnail/m.trnsz.com/hoYQMoAJAXVCXDJCIoVpYQvN?width=800&height=800","twitter":"","linkedIn":"","github":"https://github.com/cjdelisle","mastodon":"","picture_valid":""},{"name":"Philippe Honigman","title":"Tribute, OrgTech","picture":"","twitter":"https://twitter.com/phil_h","linkedIn":"https://www.linkedin.com/in/honigman/","github":"https://github.com/PhilH","mastodon":"@PhilH@mastodon.social","picture_valid":"OK"},{"name":"Amira Bouguera","title":"HellHound, Consensys","picture":"https://media.licdn.com/dms/image/C4D03AQHdI5HlgGLoHQ/profile-displayphoto-shrink_200_200/0?e=1582156800&v=beta&t=KXpAoZmsN8hE8QSrpSWEtAxko486di4ellnsywDdq7o","twitter":"https://twitter.com/amirabouguera","linkedIn":"https://www.linkedin.com/in/amira-bouguera-282025a6/","github":"","mastodon":"","picture_valid":"OK"},{"name":"Nicholas G.S. Saul","title":"Doctorant Sciences Po","picture":"https://pbs.twimg.com/profile_images/987781662158262272/0LoF0ZQF_400x400.jpg","twitter":"https://twitter.com/NicholasGSST","linkedIn":"https://www.linkedin.com/in/nicholas-g-s-saul-88414532/","github":"","mastodon":"","picture_valid":"OK"},{"name":"Liliana Carrillo","title":"P2P Foundation","picture":"/img/festival/picture_liliana_carillo.png","twitter":"https://twitter.com/lilicarrillof","linkedIn":"https://www.linkedin.com/in/carrilloliliana/","github":"","mastodon":"","picture_valid":"OK"},{"name":"Rémi Crussière","title":"Film PROTOCOLE","picture":"https://pbs.twimg.com/profile_images/1847499988/LTB_4062_400x400.JPG","twitter":"https://twitter.com/Remi_Crus","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Weedcoder","title":"Cypherpunk Mouvement","picture":"https://pbs.twimg.com/profile_images/1156519867358560256/3UVj7mmU_400x400.jpg","twitter":"https://twitter.com/weedcoder","linkedIn":"","github":"","mastodon":"","picture_valid":"OK"},{"name":"Matthew Schutte","title":"Holochain","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Madeleine Canavesio","title":"Mobicoop","picture":"https://pro.mobicoop.fr/wp-content/uploads/2019/07/MadeleineC-80x80-150x150.png","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Virgile Deville","title":"Open Source Politics","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Nelson Melina","title":"Aeternity","picture":"https://pbs.twimg.com/profile_images/1018803749907726336/yR0pRDKZ_400x400.jpg","twitter":"https://twitter.com/DaLN","linkedIn":"","github":"","mastodon":"","picture_valid":"OK"},{"name":"Manel Ruiz","title":"Aeternity","picture":"/img/festival/picture_manel_ruiz.jpg","twitter":"https://twitter.com/manelomanu","linkedIn":"","github":"","mastodon":"","picture_valid":"OK"},{"name":"Pierre Noizat","title":"Paymium","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Brian O'Hagan","title":"Coinhouse","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Duc Ha Duong","title":"Officience","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""},{"name":"Cédric Ringebach","title":"La Fresque pour le Climat","picture":"","twitter":"","linkedIn":"","github":"","mastodon":"","picture_valid":""}]
