+++ pre-build.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 set -e
 
-node ./scripts/generate-fest.js
-node ./scripts/generate-speakers.js
+node ./scripts/gsheetstojson.js '1z5GaDeDu2Ei_pCEFpFWSS53NZSmseBdDXuadUPh9qWE' 1547075518 'data/festival/events' 'festival'
+node ./scripts/gsheetstojson.js '12vIWrL2DnrfRlEZ23PFwv4sJO-rGtbD95WWsjirNqTk' 1242995252 'data/festival/speakers'
