+++ themes/p2pparis/layouts/partials/section_schedule.html
@@ -1,3 +1,5 @@
+{{ $festival_events := index .Site.Data.festival (print "events_" .Lang) }}
+
 <section class="page-section bg-light section-schedule" >
   <div class="container" id="schedule">
     <div class="row justify-content-center">
@@ -10,7 +12,7 @@
 
     <ul class="nav nav-pills nav-fill s-nav" id="nav-schedule">
 
-        {{ range (index .Site.Data.festival .Lang) }}
+        {{ range $festival_events }}
 
         {{- $date := time .date -}}
         {{- $month_num := printf "%d" $date.Month -}}
@@ -54,7 +56,7 @@
 
     <div class="s-content">
 
-      {{ range (index .Site.Data.festival .Lang) }}
+      {{ range $festival_events }}
       <div class="s-day" id="c-day-{{ .id }}">
         <div class="s-day-title text-center">
             <div class="pill">{{ i18n "day_long" . }}  {{ .id }}</div>
